Reverts "Changes product price to string"
commit reverted 1f99575c063ca9d0df32109bfc7ef464d95e17b1.
The last commit solved the problem, but the real problem
it wasn't in the code or in mongoose, the problem it was
a bug in the tool used to test the resquest.
It is clear after some test in Insomnia and Chrome
resquests.